src/net/xaethos/android/halparser/HALParser.java
package net.xaethos.android.halparser;

import java.net.URI;

public class HALParser implements HALEnclosure
{
    private final URI mURI;

    public HALParser(URI baseURI) {
        if (!baseURI.isAbsolute()) throw new IllegalArgumentException("Base URI must be absolute");
        mURI = baseURI;
    }

    public HALParser(String baseURI) {
        this(URI.create(baseURI));
    }

    @Override
    public URI getBaseURI() {
        return mURI;
    }

    @Override
    public HALEnclosure getEnclosure() {
        return null;
    }

}
